forecaster/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
supabase==2.11.0
python-dotenv==1.0.1
prophet==1.1.6
cmdstanpy==1.2.4
pandas==2.2.3
